requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client for API calls
httpx==0.25.2

# Data validation and parsing
pydantic==2.5.0

# Environment variables
python-dotenv==1.0.0

# CORS middleware (included with FastAPI but explicit for clarity)
# fastapi already includes starlette which has CORS

# Optional: For production deployment
gunicorn==21.2.0

# Optional: For enhanced logging
structlog==23.2.0